requirements-synth.txt
# Extra deps for oled_midi_synth.py (MIDI in + synth out)
# System package (once): sudo apt-get install -y libasound2-dev alsa-utils
python-rtmidi>=1.5.0
mido>=1.3.0
